remove redirect from modal
