+++ lib/quadblockquiz_web/live/live_helpers.ex
@@ -56,6 +56,7 @@ defmodule QuadblockquizWeb.LiveHelpers do
 
   defp hide_modal(js \\ %JS{}) do
     js
+    |> JS.push("unpause")
     |> JS.hide(to: "#modal", transition: "fade-out")
     |> JS.hide(to: "#modal-content", transition: "fade-out-scale")
   end
