+++ lib/quadblockquiz_web/templates/layout/app.html.heex
@@ -1,5 +1,7 @@
 <main role="main" class="container">
-  <p class="alert alert-info" role="alert"><%= Phoenix.Flash.get(@flash, :info) %></p>
-  <p class="alert alert-danger" role="alert"><%= Phoenix.Flash.get(@flash, :error) %></p>
+  <div :if={map_size(@flash) > 0}>
+    <p class="alert alert-info" role="alert"><%= Phoenix.Flash.get(@flash, :info) %></p>
+    <p class="alert alert-danger" role="alert"><%= Phoenix.Flash.get(@flash, :error) %></p>
+  </div>
   <%= @inner_content %>
 </main>
