+++ lib/quadblockquiz_web/templates/layout/live.html.heex
@@ -1,9 +1,22 @@
 <main role="main" class="container">
-  <p class="alert alert-info" role="alert" phx-click="lv:clear-flash" phx-value-key="info">
+  <% show_flash? = map_size(@flash) > 0 %>
+  <p
+    :if={show_flash?}
+    class="alert alert-info"
+    role="alert"
+    phx-click="lv:clear-flash"
+    phx-value-key="info"
+  >
     <%= live_flash(@flash, :info) %>
   </p>
 
-  <p class="alert alert-danger" role="alert" phx-click="lv:clear-flash" phx-value-key="error">
+  <p
+    :if={show_flash?}
+    class="alert alert-danger"
+    role="alert"
+    phx-click="lv:clear-flash"
+    phx-value-key="error"
+  >
     <%= live_flash(@flash, :error) %>
   </p>
